+++ client/Packages/com.beamable/Editor/BeamCli/Commands/BeamProjectBuild.cs
@@ -14,6 +14,8 @@ namespace Beamable.Editor.BeamCli.Commands
         public bool exactIds;
         /// <summary>A message to send to the running service when it is terminated</summary>
         public string stopReason;
+        /// <summary>Maximum number of parallel services builds</summary>
+        public int maxParallelCount;
         /// <summary>Serializes the arguments for command line usage.</summary>
         public virtual string Serialize()
         {
@@ -42,6 +44,11 @@ namespace Beamable.Editor.BeamCli.Commands
             if ((this.stopReason != default(string)))
             {
                 genBeamCommandArgs.Add(("--stop-reason=" + this.stopReason));
+            }
+            // If the maxParallelCount value was not default, then add it to the list of args.
+            if ((this.maxParallelCount != default(int)))
+            {
+                genBeamCommandArgs.Add(("--max-parallel-count=" + this.maxParallelCount));
             }
             string genBeamCommandStr = "";
             // Join all the args with spaces